move common stuff to one module
